+++ packages/core/client/src/modules/blocks/data-blocks/table/hooks/useTableBlockProps.tsx
@@ -66,7 +66,6 @@ export const useTableBlockProps = () => {
   }, [field, service?.data, service?.loading, tableBlockContextBasicValue.field?.data?.selectedRowKeys]);
 
   return {
-    optimizeTextCellRender: true,
     value: data,
     childrenColumnName: tableBlockContextBasicValue.childrenColumnName,
     loading: service?.loading,