+++ packages/core/client/src/modules/fields/component/DatePicker/datePickerComponentFieldSettings.tsx
@@ -9,8 +9,9 @@
 
 import { useFieldSchema } from '@formily/react';
 import { SchemaSettings } from '../../../../application/schema-settings/SchemaSettings';
-import { useColumnSchema } from '../../../../schema-component/antd/table-v2/Table.Column.Decorator';
 import { SchemaSettingsDateFormat } from '../../../../schema-settings/SchemaSettingsDateFormat';
+import { SchemaSettingsDateRange } from '../../../../schema-settings/SchemaSettingsDateRange';
+import { useColumnSchema } from '../../../../schema-component/antd/table-v2/Table.Column.Decorator';
 import { ellipsisSettingsItem, enableLinkSettingsItem, openModeSettingsItem } from '../Input/inputComponentSettings';
 
 export const datePickerComponentFieldSettings = new SchemaSettings({
@@ -26,6 +27,24 @@ export const datePickerComponentFieldSettings = new SchemaSettings({
         return {
           fieldSchema,
         };
+      },
+    },
+    {
+      name: 'dateScopeSelect',
+      Component: SchemaSettingsDateRange as any,
+      useComponentProps() {
+        const schema = useFieldSchema();
+        const { fieldSchema: tableColumnSchema } = useColumnSchema();
+        const fieldSchema = tableColumnSchema || schema;
+        return {
+          fieldSchema,
+        };
+      },
+      useVisible() {
+        const { fieldSchema: columnSchema } = useColumnSchema();
+        const schema = useFieldSchema();
+        const fieldSchema = columnSchema || schema;
+        return !fieldSchema?.['x-read-pretty'];
       },
     },
     ellipsisSettingsItem,