+++ packages/plugins/@nocobase/plugin-notification-in-app-message/src/server/defineMyInAppChannels.ts
@@ -73,6 +73,7 @@ export default function defineMyInAppChannels({ app }: { app: Application }) {
               SELECT  messages.${messagesFieldName.channelName}
               FROM ${messagesTableName} AS messages
               WHERE messages.${messagesFieldName.status} = '${status}'
+              AND messages.${messagesFieldName.userId} = ${userId}
           )`);
             return { name: { [Op.in]: sql } };
           };