+++ packages/plugins/@nocobase/plugin-workflow-aggregate/src/server/AggregateInstruction.ts
@@ -43,7 +43,7 @@ export default class extends Instruction {
     const result = await repo.aggregate({
       ...options,
       method: aggregators[aggregator],
-      // transaction: processor.transaction,
+      transaction: this.workflow.useDataSourceTransaction(dataSourceName, processor.transaction),
     });
 
     return {